api/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
groq>=0.13.1
pydantic>=2.11.0
python-dotenv>=1.0.1
pandas>=2.2.3
requests>=2.32.3
beautifulsoup4>=4.12.3
python-multipart>=0.0.20
httpx>=0.28.0
